Add 'html' and 'assets' cargo features
That way, users of this library can remove support for highlighting
that is not needed for them.

That way, binaries will be smaller (e.g. no static asset data), and
compile times will be reduced as well.

This commit is motivated by the `tokio-cassandra` shell, which will
use only use two syntaxes, YAML + SQL/CQL, and a handful of themes.