Add "yaml-load" feature flag
Default on.
If off, there will be no usage of yaml-rust, which turns out to
be problematic if you are already using it with different feature
flags (i.e. linked hashmap vs. btreemap).

This also means the only way to create a syntax set is using the
functions from `dumps`, and it might be nice to have one of these
directly in the SyntaxSet type itself.