Improve regex rewriting for nonewlines mode (#76)
Before, a regex like `[\n]` would be rewritten as `[\z]`, which just
matches a literal `z` (because anchors don't work in character classes).

The new logic parses a limited amount of regex syntax to understand
escapes and character classes. As a result, it does a better job than
before. Note that it's still just a best effort.