Correct order of iterating with_prototypes.
When I first wrote that code I assumed that prototypes were used from top to
bottom but I tested it and it's actually bottom to top.

This reduces the number of failing ASP tests but increases the number of
failing PHP tests. I expect the new PHP failures might just be exposing
another bug.