Greedily merge consecutive spans on the same line
This merges two or more consecutive spans of the same color
that are on the same line. If the span on the right is
a whitespace, and the background colours are identical,
a merger happens as well.

This allows us to save about 50% of character count in the HTML
output.