Recover from catastrophic backtracking
Oniguruma 6.8.0 enabled USE_TRY_IN_MATCH_LIMIT by default, with a limit
of 10_000_000. That means after that number of failed matches, it will
stop trying to match and return an error instead.

We can gracefully handle that error by using the onig crate's new
`search_with_param` method which returns a Result instead of panicking.

I tried this on the file from #64 and the parse time goes from 10
seconds to 200 ms.