Use line-local state instead of StateLevel and line number
This makes the code simpler and it's actually not needed to store more
than one of these states (as it was before), because once we consume a
character, we don't need any earlier state anymore.

It also has the benefit of allowing caching of the parse state and
re-parsing the same line multiple times, which would not have made sense
with the line number before.