Add SyntaxSetBuilder to get rid of manual linking
Previously, linking had to be done manually using `link_syntaxes` in
order for the syntaxes to work.

This change introduces a SyntaxSetBuilder that contains all the methods
to add/load syntaxes. When calling `build`, the syntaxes are linked and
you get a `SyntaxSet`.

This has the nice side effect of separating all the methods that mutate
to a separate type and making SyntaxSet immutable.