Switch HTML helpers to newlines syntaxes
So that they are faster and more correct. Also refactors them to not allocate
intermediate strings and share more code.

Also renames the methods which had silent requirements changes so that people
upgrading get compile errors and hopefully then look at the release notes.

There are two other functionality changes: the position of the newline
characters in the resulting HTML changed, in a way I think is reasonable and
doesn't change the rendered output. And the `start_coloured_html_snippet`
function got a semicolon added and now returns the background.

Signed-off-by: Tristan Hume <tris.hume@gmail.com>