Add support for loading metadata files
This adds initial support for loading metadata from `.tmPreferences`
files. The main focus of this patch is loading the patterns used to
determine indentation. I expect to follow up with support for comment
strings, with support for other features possible in the future.

This is all gated behind a new `metadata` feature. If enabled, a
top-level `Metadata` object will be included in `SyntaxSet`; this object
can be queried for the metadata available for a given scope.

This patch does not attempt to implement any actual indentation logic;
that is left up to the consumer.