Added `RangedHighlightIterator`
There're situations where it is useful to know the range of a
highlighted substring in the source string. For example, to just
paint existing text instead of rebuilding it (i.e in a text editor).

This implements a `RangedHighlightIterator` in a backwards
compatible manner, so that the previous `HighlightIterator`
still exists but uses the same implementation.