Fix rewriting of "newlines" mode regexes
Always adding `(?m)` for the entire regex meant that `.` also changed meaning,
which is not what we want. The safer option is to use `(?m:$)` for `$` only.

That also means we don't have to bother with `\A`. But we do need to parse
look-behinds because we can't use `(?m:$)` in it.