Revert breaking API change
This changed a public API in a breaking way which would require a major
version release. It's also not necessary and the current API is in
idiomatic Rust style as far as I know.

If the existing API makes it harder to figure out borrow checker issues
you may be able to solve your problem by changing things around so you
can move the HTML generator, or use mem::swap to swap in an empty one.