html: allow a configurable prefix on CSS class names.
CSS class names are a global namespace. Without some kind of namespace
partitioning, it's entirely possible to have class name collisions and
all sorts of chaos as different sources of CSS stomp on each other.

Hence, add a new `ClassStyle`, `SpacedPrefixed`, which tells syntact
to add a configurable prefix to all of the class names that it uses.

This patch breaks the API. Some of the breakage is not strictly
necessary (e.g., `ClassedHTMLGenerator::new` could stay as it is, and
a `new_with_style` method added). However, some is enforced:
`ClassStyle` can no longer be `Copy` because it can now contain a
`String`, and so `tokens_to_classed_spans` must now accept a
reference. Because this is a breaking API change anyway, I took the
opportunity to add `ClassStyle` parameters where relevant, to make the
parametricity of `ClassedHTMLGenerator` and `css_for_theme` clear.