Fix handling of newlines in classed HTML
The examples used the `_newlines` syntax without providing newlines
leading to broken highlighting (#303). The function for producing
classed HTML added newlines for you so I deprecated that one and
added a new one which doesn't, so that I could change the examples to
use the `_newlines` variant in the proper way.