Make 'plist' dependency optional via new 'plist-load' feature
plist brings in several dependencies, most notably the heavy xml-rs dependency.
So if you don't need to parse theme files, it is nice to be able to skip
this dependency.

For example it brings down the number of dependencies to build from 68 to 52
when building examples/synhtml.rs as a regular app. And build time shrinks
proportionally. Try yourself with:

    cp examples/synhtml.rs src/main.rs # temporarily side-step dev-dependencies
    cargo build --no-default-features --features parsing,html,assets,dump-load,regex-onig