Fix all clippy::option-as-ref-deref warnings
We currently just have one of this.

`line_comment` is of type `Option<String>` and we want an `Option<&str>`.
We currently convert first to `Option<&String>` and then `map` to `.as_str()`.
But since `String` implements `Deref` like this:

    impl ops::Deref for String {
        type Target = str;

        #[inline]
        fn deref(&self) -> &str {
            unsafe { str::from_utf8_unchecked(&self.vec) }
        }
    }

we can simplify to just use `.as_deref()`, as recommended by clippy.