Increase MSRV instead of putting upper bound on bitflags
It is getting unreasonably much work to stay at MSRV 1.42.

bat recently bumped nix to 0.23.0 which requires bitflags 1.3.1 or later.
See https://github.com/sharkdp/bat/pull/1874. So as of now, bat can't build
with syntect master.

It will cause unnecssary headaches for bat and other downstream projects if we
stay at MSRV 1.42. So remove upper bound on bitflags and bump MSRV to 1.46
instead and only put a lower bound (same as we had before) on bitflags.