Lazy-load syntax Contexts
This enables making `SyntaxSet::load_defaults_newlines()` *much* faster.
Intuitively it might seem like this change is enough, but it is also necessary
to stop compressing the "outer" data in `assets/default_newlines.packdump`.
Since the inner, lazy-loaded data is already compressed, compressing the data
another time only makes the data CPU-intensive to decompress, without any gains
in reduced size of the data.

The final step in lazy-loading will be done in a separate commit. Here are an
overview of what kind of improvements are at stake:

| Version              | `SyntaxSet::load_defaults_newlines()` time |
| -------------------- | -----------------------------------------  |
| v4.6.0               | 35.3 ms                                    |
| This commit          | 24.0 ms                                    |
| No outer compression |  1.3 ms (!)                                |

Lazy-loading syntaxes is very fast (in the sub-ms range), so time-to-parse is
practically unchanged by this commit.