Silence criterion HTML report generation warning
Without this fix, the following warning is printed when doing `cargo bench`.

WARNING: HTML report generation will become a non-default optional feature in
Criterion.rs 0.4.0. This feature is being moved to cargo-criterion
(https://github.com/bheisler/cargo-criterion) and will be optional in a future
version of Criterion.rs. To silence this warning, either switch to
cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.