benches: Use exact same names as in v4.6.0
When doing #380 I were not insightful enough to realize that it will be useful
to have the exact same names as before. This is useful when using the
`--baseline` feature of Criterion to compare current code with old code.

After this commit, we can do this:

```bash
git checkout v4.6.0
cargo bench 'highlight/"jquery.js"' -- --save-baseline v4.6.0
git checkout origin/master
cargo bench 'highlight/"jquery.js"' -- --baseline v4.6.0
```

Without this commit, the name of the benchmark on `master` is
`'highlight/jquery.js'`.