Add Criterion benchmark test that includes loading assets
This makes it easy to measure the impact lazy-loading will have on a complete
syntect "pipeline".

This commit is designed to be cherry-pickable to the v4.6.0 tag to get a
Criterion `--baseline` for the lazy-loading changes. An upcoming commit will
make existing benches make use of the helper methods that this commit adds.