Remove `plist-load` feature (but keep new code structure)
It was not enough to make the `default` feature depend on it. Putting code
behind a feature is in practice a semver breakage. See
https://github.com/sharkdp/bat/issues/1991#issuecomment-1001203088 and #402.

It was added in #345 after doing some refactoring. We can keep the refactoring,
but we choose to remove the feature for now. Another option would be to make the
feature opt-out rather than opt-in and rename it to `no-plist-load` or similar.
That would be too unergonomic and confusing however, so for now we just remove
it. We might add it back for v5.0.0.