Replace direct panics with Errors
Even though there are still `.unwrap()`s and `.expects()` left, this
change should cover the most common error code paths.

If we find a new case of an error that we want to propagate, it is
generally straightforward to deprecate the old function and add a new
function that the old function calls and `.unwrap()`.