Use `.txt` file for public API snapshot instead of `.snap` (from `insta`)
So that showing the diff in GitHub does not become more or less
unreadable because of faulty syntax highlighting.

The `public-api` crate now has (optional) support for built-in snapshot
testing that we use instead of `insta`. But note that the excellent
diffing engine of `insta` (from the crate `similar`) is still used.

We get two new indirect `dev-dependencies`:
* `similar-asserts` - from the same author as `similar` and `insta`, so
 we can trust it.
* `snapshot-testing` - new but tiny and easily auditable. (And the
author happens to be me.)

If the `public_api` test fails, instructions on how to update are printed:

    thread 'public_api' panicked at tests/public_api.rs:20:16:
    assertion failed: `(left == right)`:

    help: run with env var `UPDATE_SNAPSHOTS=yes` to update snapshots

    ...